Complete overhaul of app logic
I don't know why I bother, tbh. With a centralised module it is easier to have a complete view of the network to power more advanced features in the app. The three controllers 'RoutesController', 'StopsController' and 'VehiclesController' were combined in a single Controller, temporarily named "TheManager" - will change of course.