Today I discovered the Singleton
And it is incredible. Essentially is the best of every world. With a Singleton, it is possible to have the same instance of a class shared across the whole application. If you change the value in one place, it updates everywhere. It is the best, so I converted Appstate, Analytics and CarrisAuthentication to this new format. I feel it is cleaner due to the _inexistent_ plumbing needed for the same functionality.